Switch framerate limiter to using doubles

+++ Descent3/init.cpp
@@ -139,7 +139,7 @@ ubyte Use_motion_blur = 0;
 char Base_directory[_MAX_PATH];
 char* User_directory;
 
-extern float Min_allowed_frametime;
+extern double Min_allowed_frametime;
 
 extern bool Mem_low_memory_mode;
 extern bool Render_powerup_sparkles;
@@ -302,20 +302,20 @@ void PreInitD3Systems()
 	int iframelmtarg = FindArg("-limitframe");
 	if(iframelmtarg)
 	{
-		Min_allowed_frametime = atoi(GameArgs[iframelmtarg+1]) / 1000.f;
+		Min_allowed_frametime = atoi(GameArgs[iframelmtarg+1]) / 1000.0;
 		mprintf((0,"Using %d as a minimum frametime\n",Min_allowed_frametime));
 	}
 	else
 	{
 		if(FindArg("-dedicated"))
-			Min_allowed_frametime=1.f/30;
+			Min_allowed_frametime=1./30;
 		else
 			Min_allowed_frametime=0;
 	}
 	iframelmtarg = FindArg("-framecap");
 	if(iframelmtarg)
 	{
-		Min_allowed_frametime = ((float)1.0/(float)atoi(GameArgs[iframelmtarg+1]));
+		Min_allowed_frametime = (1.0/atoi(GameArgs[iframelmtarg+1]));
 		mprintf((0,"Using %f as a minimum frametime\n",Min_allowed_frametime));
 	}
 	else
